Convert all approval-screen option pickers to bottom sheets
Extract the bottom-sheet picker from RememberMyChoice into a generic
OptionBottomSheetPicker<T> and reuse it for:
  - Allow / Deny / Ask on each per-app permission row
  - "For this relay only / For all relays" on relay-auth approvals
    (BunkerRelayAuthScreen, BunkerMultiEventHomeScreen,
    IntentMultiEventHomeScreen)
  - "For this method only / For all methods" on encrypt/decrypt
    approvals (EncryptDecryptData, 2 sites)
